Add files via upload
Write a Python Monster class.  Monster should have the following:

an  __init__ method that takes the name and the size of the monster and creates and sets instance variables with these values
an attack method that takes the name of the location the monster attacks and prints out a message like
Godzilla attacks Tokyo

getters for the name and size of the monster
a reasonable __str__ method
an __eq__ method that overloadss the == operator so that two monsters are equal if and only if they have the same name and the same size
an __add__ method that returns a new Monster whose name is the names of the two original monsters concatenated (for example, GodzillaKong) and size is the combined size of the two original monsters
Write test code that demonstrates that all your methods work correctly.  Turn in your .py file and a copy of the output from a sample run of your test code.